Clean up client and server a little.
