+++ server/run_controller_cli.py
@@ -3,12 +3,11 @@
 import argparse
 import asyncio
 import logging
-import time
 import os
-from threading import Thread
 
 from aioconsole import ainput
-
+from flask import Flask
+from flask_socketio import SocketIO
 from joycontrol import logging_default as log, utils
 from joycontrol.command_line_interface import ControllerCLI
 from joycontrol.controller import Controller
@@ -17,8 +16,7 @@ from joycontrol.memory import FlashMemory
 from joycontrol.protocol import controller_protocol_factory
 from joycontrol.server import create_hid_server
 
-from flask import Flask
-from flask_socketio import SocketIO
+from server.switchremoteplay.controller import SwitchController
 
 logger = logging.getLogger(__name__)
 
@@ -56,324 +54,269 @@ Options:
 
 
 async def test_controller_buttons(controller_state: ControllerState):
-    """
-    Example controller script.
-    Navigates to the "Test Controller Buttons" menu and presses all buttons.
-    """
-    if controller_state.get_controller() != Controller.PRO_CONTROLLER:
-        raise ValueError('This script only works with the Pro Controller!')
+	"""
+	Example controller script.
+	Navigates to the "Test Controller Buttons" menu and presses all buttons.
+	"""
+	if controller_state.get_controller() != Controller.PRO_CONTROLLER:
+		raise ValueError('This script only works with the Pro Controller!')
 
-    # waits until controller is fully connected
-    await controller_state.connect()
+	# waits until controller is fully connected
+	await controller_state.connect()
 
-    await ainput(prompt='Make sure the Switch is in the Home menu and press <enter> to continue.')
+	await ainput(prompt='Make sure the Switch is in the Home menu and press <enter> to continue.')
 
-    """
-    # We assume we are in the "Change Grip/Order" menu of the switch
-    await button_push(controller_state, 'home')
+	"""
+	# We assume we are in the "Change Grip/Order" menu of the switch
+	await button_push(controller_state, 'home')
 
-    # wait for the animation
-    await asyncio.sleep(1)
-    """
+	# wait for the animation
+	await asyncio.sleep(1)
+	"""
 
-    # Goto settings
-    await button_push(controller_state, 'down', sec=1)
-    await button_push(controller_state, 'right', sec=2)
-    await asyncio.sleep(0.3)
-    await button_push(controller_state, 'left')
-    await asyncio.sleep(0.3)
-    await button_push(controller_state, 'a')
-    await asyncio.sleep(0.3)
+	# Goto settings
+	await button_push(controller_state, 'down', sec=1)
+	await button_push(controller_state, 'right', sec=2)
+	await asyncio.sleep(0.3)
+	await button_push(controller_state, 'left')
+	await asyncio.sleep(0.3)
+	await button_push(controller_state, 'a')
+	await asyncio.sleep(0.3)
 
-    # go all the way down
-    await button_push(controller_state, 'down', sec=4)
-    await asyncio.sleep(0.3)
+	# go all the way down
+	await button_push(controller_state, 'down', sec=4)
+	await asyncio.sleep(0.3)
 
-    # goto "Controllers and Sensors" menu
-    for _ in range(2):
-        await button_push(controller_state, 'up')
-        await asyncio.sleep(0.3)
-    await button_push(controller_state, 'right')
-    await asyncio.sleep(0.3)
+	# goto "Controllers and Sensors" menu
+	for _ in range(2):
+		await button_push(controller_state, 'up')
+		await asyncio.sleep(0.3)
+	await button_push(controller_state, 'right')
+	await asyncio.sleep(0.3)
 
-    # go all the way down
-    await button_push(controller_state, 'down', sec=3)
-    await asyncio.sleep(0.3)
+	# go all the way down
+	await button_push(controller_state, 'down', sec=3)
+	await asyncio.sleep(0.3)
 
-    # goto "Test Input Devices" menu
-    await button_push(controller_state, 'up')
-    await asyncio.sleep(0.3)
-    await button_push(controller_state, 'a')
-    await asyncio.sleep(0.3)
+	# goto "Test Input Devices" menu
+	await button_push(controller_state, 'up')
+	await asyncio.sleep(0.3)
+	await button_push(controller_state, 'a')
+	await asyncio.sleep(0.3)
 
-    # goto "Test Controller Buttons" menu
-    await button_push(controller_state, 'a')
-    await asyncio.sleep(0.3)
+	# goto "Test Controller Buttons" menu
+	await button_push(controller_state, 'a')
+	await asyncio.sleep(0.3)
 
-    # push all buttons except home and capture
-    button_list = controller_state.button_state.get_available_buttons()
-    if 'capture' in button_list:
-        button_list.remove('capture')
-    if 'home' in button_list:
-        button_list.remove('home')
+	# push all buttons except home and capture
+	button_list = controller_state.button_state.get_available_buttons()
+	if 'capture' in button_list:
+		button_list.remove('capture')
+	if 'home' in button_list:
+		button_list.remove('home')
 
-    user_input = asyncio.ensure_future(
-        ainput(prompt='Pressing all buttons... Press <enter> to stop.')
-    )
+	user_input = asyncio.ensure_future(
+		ainput(prompt='Pressing all buttons... Press <enter> to stop.')
+	)
 
-    # push all buttons consecutively until user input
-    while not user_input.done():
-        for button in button_list:
-            await button_push(controller_state, button)
-            await asyncio.sleep(0.1)
+	# push all buttons consecutively until user input
+	while not user_input.done():
+		for button in button_list:
+			await button_push(controller_state, button)
+			await asyncio.sleep(0.1)
 
-            if user_input.done():
-                break
+			if user_input.done():
+				break
 
-    # await future to trigger exceptions in case something went wrong
-    await user_input
+	# await future to trigger exceptions in case something went wrong
+	await user_input
 
-    # go back to home
-    await button_push(controller_state, 'home')
+	# go back to home
+	await button_push(controller_state, 'home')
 
 
 async def set_amiibo(controller_state, file_path):
-    """
-    Sets nfc content of the controller state to contents of the given file.
-    :param controller_state: Emulated controller state
-    :param file_path: Path to amiibo dump file
-    """
-    loop = asyncio.get_event_loop()
+	"""
+	Sets nfc content of the controller state to contents of the given file.
+	:param controller_state: Emulated controller state
+	:param file_path: Path to amiibo dump file
+	"""
+	loop = asyncio.get_event_loop()
 
-    with open(file_path, 'rb') as amiibo_file:
-        content = await loop.run_in_executor(None, amiibo_file.read)
-        controller_state.set_nfc(content)
+	with open(file_path, 'rb') as amiibo_file:
+		content = await loop.run_in_executor(None, amiibo_file.read)
+		controller_state.set_nfc(content)
 
 
 async def mash_button(controller_state, button, interval):
-    # waits until controller is fully connected
-    await controller_state.connect()
+	# waits until controller is fully connected
+	await controller_state.connect()
 
-    if button not in controller_state.button_state.get_available_buttons():
-        raise ValueError(f'Button {button} does not exist on {controller_state.get_controller()}')
+	if button not in controller_state.button_state.get_available_buttons():
+		raise ValueError(f'Button {button} does not exist on {controller_state.get_controller()}')
 
-    user_input = asyncio.ensure_future(
-        ainput(prompt=f'Pressing the {button} button every {interval} seconds... Press <enter> to stop.')
-    )
-    # push a button repeatedly until user input
-    while not user_input.done():
-        await button_push(controller_state, button)
-        await asyncio.sleep(float(interval))
+	user_input = asyncio.ensure_future(
+		ainput(prompt=f'Pressing the {button} button every {interval} seconds... Press <enter> to stop.')
+	)
+	# push a button repeatedly until user input
+	while not user_input.done():
+		await button_push(controller_state, button)
+		await asyncio.sleep(float(interval))
 
-    # await future to trigger exceptions in case something went wrong
-    await user_input
+	# await future to trigger exceptions in case something went wrong
+	await user_input
 
 
 async def _main(args):
-    # parse the spi flash
-    if args.spi_flash:
-        with open(args.spi_flash, 'rb') as spi_flash_file:
-            spi_flash = FlashMemory(spi_flash_file.read())
-    else:
-        # Create memory containing default controller stick calibration
-        spi_flash = FlashMemory()
+	# parse the spi flash
+	if args.spi_flash:
+		with open(args.spi_flash, 'rb') as spi_flash_file:
+			spi_flash = FlashMemory(spi_flash_file.read())
+	else:
+		# Create memory containing default controller stick calibration
+		spi_flash = FlashMemory()
 
-    # Get controller name to emulate from arguments
-    controller = Controller.from_arg(args.controller)
+	# Get controller name to emulate from arguments
+	controller = Controller.from_arg(args.controller)
 
-    with utils.get_output(path=args.log, default=None) as capture_file:
-        print(f"capture_file: {capture_file}")
-        factory = controller_protocol_factory(controller, spi_flash=spi_flash)
-        ctl_psm, itr_psm = 17, 19
-        transport, protocol = await create_hid_server(factory, reconnect_bt_addr=args.reconnect_bt_addr,
-                                                      ctl_psm=ctl_psm,
-                                                      itr_psm=itr_psm, capture_file=capture_file,
-                                                      device_id=args.device_id)
+	with utils.get_output(path=args.log, default=None) as capture_file:
+		print(f"capture_file: {capture_file}")
+		factory = controller_protocol_factory(controller, spi_flash=spi_flash)
+		ctl_psm, itr_psm = 17, 19
+		transport, protocol = await create_hid_server(factory, reconnect_bt_addr=args.reconnect_bt_addr,
+		                                              ctl_psm=ctl_psm,
+		                                              itr_psm=itr_psm, capture_file=capture_file,
+		                                              device_id=args.device_id)
 
-        controller_state = protocol.get_controller_state()
+		controller_state = protocol.get_controller_state()
 
-        # Create command line interface and add some extra commands
-        cli = ControllerCLI(controller_state)
+		# Create command line interface and add some extra commands
+		cli = ControllerCLI(controller_state)
 
-        # Wrap the script so we can pass the controller state. The doc string will be printed when calling 'help'
-        async def _run_test_controller_buttons():
-            """
-            test_buttons - Navigates to the "Test Controller Buttons" menu and presses all buttons.
-            """
-            await test_controller_buttons(controller_state)
+		# Wrap the script so we can pass the controller state. The doc string will be printed when calling 'help'
+		async def _run_test_controller_buttons():
+			"""
+			test_buttons - Navigates to the "Test Controller Buttons" menu and presses all buttons.
+			"""
+			await test_controller_buttons(controller_state)
 
-        # add the script from above
-        cli.add_command('test_buttons', _run_test_controller_buttons)
+		# add the script from above
+		cli.add_command('test_buttons', _run_test_controller_buttons)
 
-        # Mash a button command
-        async def call_mash_button(*args):
-            """
-            mash - Mash a specified button at a set interval
+		# Mash a button command
+		async def call_mash_button(*args):
+			"""
+			mash - Mash a specified button at a set interval
 
-            Usage:
-                mash <button> <interval>
-            """
-            if not len(args) == 2:
-                raise ValueError('"mash_button" command requires a button and interval as arguments!')
+			Usage:
+				mash <button> <interval>
+			"""
+			if not len(args) == 2:
+				raise ValueError('"mash_button" command requires a button and interval as arguments!')
 
-            button, interval = args
-            await mash_button(controller_state, button, interval)
+			button, interval = args
+			await mash_button(controller_state, button, interval)
 
-        # add the script from above
-        cli.add_command('mash', call_mash_button)
+		# add the script from above
+		cli.add_command('mash', call_mash_button)
 
-        # Create amiibo command
-        async def amiibo(*args):
-            """
-            amiibo - Sets nfc content
+		# Create amiibo command
+		async def amiibo(*args):
+			"""
+			amiibo - Sets nfc content
 
-            Usage:
-                amiibo <file_name>          Set controller state NFC content to file
-                amiibo remove               Remove NFC content from controller state
-            """
-            if controller_state.get_controller() == Controller.JOYCON_L:
-                raise ValueError('NFC content cannot be set for JOYCON_L')
-            elif not args:
-                raise ValueError('"amiibo" command requires file path to an nfc dump as argument!')
-            elif args[0] == 'remove':
-                controller_state.set_nfc(None)
-                print('Removed nfc content.')
-            else:
-                await set_amiibo(controller_state, args[0])
+			Usage:
+				amiibo <file_name>          Set controller state NFC content to file
+				amiibo remove               Remove NFC content from controller state
+			"""
+			if controller_state.get_controller() == Controller.JOYCON_L:
+				raise ValueError('NFC content cannot be set for JOYCON_L')
+			elif not args:
+				raise ValueError('"amiibo" command requires file path to an nfc dump as argument!')
+			elif args[0] == 'remove':
+				controller_state.set_nfc(None)
+				print('Removed nfc content.')
+			else:
+				await set_amiibo(controller_state, args[0])
 
-        # add the script from above
-        cli.add_command('amiibo', amiibo)
+		# add the script from above
+		cli.add_command('amiibo', amiibo)
 
-        try:
-            l = asyncio.get_running_loop()
-            def f():
-                start_socket_server(controller_state, l)
+		try:
+			l = asyncio.get_running_loop()
 
-            l.run_in_executor(None, f)
+			def f():
+				start_socket_server(controller_state, l)
 
-            print("Calling cli.run()")
-            await cli.run()
-        finally:
-            logger.info('Stopping communication...')
-            await transport.close()
+			l.run_in_executor(None, f)
+
+			print("Calling cli.run()")
+			await cli.run()
+		finally:
+			logger.info('Stopping communication...')
+			await transport.close()
+
 
 def start_socket_server(controller, loop):
-    print("STARTING SERVER")
-    app = Flask(__name__)
-    app.config['SECRET_KEY'] = os.getenv('SECRET_KEY') or 'asdasdasdf'
-    # print(app.config['SECRET_KEY'])
-    socketio = SocketIO(app, cors_allowed_origins='*')
-    asyncio.set_event_loop(loop)
+	print("STARTING SERVER")
+	app = Flask(__name__)
+	app.config['SECRET_KEY'] = os.getenv('SECRET_KEY') or 'asdasdasdf'
+	# print(app.config['SECRET_KEY'])
+	socketio = SocketIO(app, cors_allowed_origins='*')
+	asyncio.set_event_loop(loop)
+	controller = SwitchController(controller)
 
-    @socketio.on('connect')
-    def test_connect():
-        print("Connected")
+	@socketio.on('connect')
+	def test_connect():
+		print("Connected")
 
-    @socketio.on('disconnect')
-    def test_disconnect():
-        print("Disconnected")
+	@socketio.on('disconnect')
+	def test_disconnect():
+		print("Disconnected")
 
-    # Handle unnamed events.
-    @socketio.on('message')
-    def handle_message(message):
-        print('received message: ' + message)
+	# Handle unnamed events.
+	@socketio.on('message')
+	def handle_message(message):
+		print('received message: ' + message)
 
-    # Handle unnamed events with JSON.
-    @socketio.on('json')
-    def handle_json(json):
-        print('received json: ' + str(json))
+	# Handle unnamed events with JSON.
+	@socketio.on('json')
+	def handle_json(json):
+		print('received json: ' + str(json))
 
-    def _set_stick(stick, direction, value):
-        if direction == 'center':
-            stick.set_center()
-        elif direction == 'up':
-            stick.set_up()
-        elif direction == 'down':
-            stick.set_down()
-        elif direction == 'left':
-            stick.set_left()
-        elif direction == 'right':
-            stick.set_right()
-        elif direction in ('h', 'horizontal'):
-            if value is None:
-                raise ValueError(f'Missing value')
-            try:
-                val = int(value)
-            except ValueError:
-                raise ValueError(f'Unexpected stick value "{value}"')
-            stick.set_h(val)
-        elif direction in ('v', 'vertical'):
-            if value is None:
-                raise ValueError(f'Missing value')
-            try:
-                val = int(value)
-            except ValueError:
-                raise ValueError(f'Unexpected stick value "{value}"')
-            stick.set_v(val)
-        else:
-            raise ValueError(f'Unexpected argument "{direction}"')
+	@socketio.on('p')
+	def handle_press(command: str):
+		controller.run(command)
+		return "DONE {}".format(command)
 
-        return f'{stick.__class__.__name__} was set to ({stick.get_h()}, {stick.get_v()}).'
+	host = '0.0.0.0'
+	# 5000 is the default port.
+	port = 5000
+	print("Running at {}:{}".format(host, port))
 
+	socketio.run(app, host, port)
 
-    @socketio.on('p')
-    def handle_press(command: str):
-        logger.debug(command)
-        if command in 'lrudabxy':
-            asyncio.ensure_future(button_push(controller, command))
-        elif command.startswith('s'):
-            command = command.split(' ')
-            assert len(command) >= 3
-            if command[1] == 'l':
-                stick = controller.l_stick_state
-            else:
-                stick = controller.r_stick_state
-            direction = command[2]
-            if len(command) > 3:
-                value = command[3]
-            else:
-                value = None
-            s = _set_stick(stick, direction, value)
-            logger.debug(s)
-            asyncio.ensure_future(controller.send())
-        else:
-            command = command.split(' ')
-            assert len(command) >= 2
-            button = command[0]
-            pushed = command[1] == 'd'
-            controller.button_state.set_button(button, pushed)
-
-
-        return "DONE {}".format(command)
-
-
-    host = '0.0.0.0'
-    # 5000 is the default port.
-    port = 5000
-    print("Running at {}:{}".format(host, port))
-
-    socketio.run(app, host, port)
 
 if __name__ == '__main__':
-    # check if root
-    if not os.geteuid() == 0:
-        raise PermissionError('Script must be run as root!')
+	# check if root
+	if not os.geteuid() == 0:
+		raise PermissionError('Script must be run as root!')
 
-    # setup logging
-    #log.configure(console_level=logging.ERROR)
-    log.configure()
+	# setup logging
+	# log.configure(console_level=logging.ERROR)
+	log.configure()
 
-    parser = argparse.ArgumentParser()
-    parser.add_argument('controller', help='JOYCON_R, JOYCON_L or PRO_CONTROLLER')
-    parser.add_argument('-l', '--log')
-    parser.add_argument('-d', '--device_id')
-    parser.add_argument('--spi_flash')
-    parser.add_argument('-r', '--reconnect_bt_addr', type=str, default=None,
-                        help='The Switch console Bluetooth address, for reconnecting as an already paired controller')
-    args = parser.parse_args()
+	parser = argparse.ArgumentParser()
+	parser.add_argument('controller', help='JOYCON_R, JOYCON_L or PRO_CONTROLLER')
+	parser.add_argument('-l', '--log')
+	parser.add_argument('-d', '--device_id')
+	parser.add_argument('--spi_flash')
+	parser.add_argument('-r', '--reconnect_bt_addr', type=str, default=None,
+	                    help='The Switch console Bluetooth address, for reconnecting as an already paired controller')
+	args = parser.parse_args()
 
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(
-        _main(args)
-    )
-    print("Done __main__")
+	loop = asyncio.get_event_loop()
+	loop.run_until_complete(
+		_main(args)
+	)
+	print("Done __main__")
